Revert merge changes
