+++ db/tables/ddl/__init__.py
@@ -1,0 +1,6 @@
+from db.tables.ddl.create import *  # noqa
+from db.tables.ddl.drop import *  # noqa
+from db.tables.ddl.merge import *  # noqa
+from db.tables.ddl.move_columns import *  # noqa
+from db.tables.ddl.split import *  # noqa
+from db.tables.ddl.update import *  # noqa
