+++ db/tests/tables/ddl/test_create.py
@@ -1,0 +1,23 @@
+from db.tables import ddl as table_ddl
+from db.tests.types import fixtures
+
+
+# We need to set these variables when the file loads, or pytest can't
+# properly detect the fixtures.  Importing them directly results in a
+# flake8 unused import error, and a bunch of flake8 F811 errors.
+engine_with_types = fixtures.engine_with_types
+engine_email_type = fixtures.engine_email_type
+temporary_testing_schema = fixtures.temporary_testing_schema
+
+
+def test_table_creation_doesnt_reuse_defaults(engine_with_schema):
+    column_list = []
+    engine, schema = engine_with_schema
+    t1 = table_ddl.create_mathesar_table("t1", schema, column_list, engine)
+    t2 = table_ddl.create_mathesar_table("t2", schema, column_list, engine)
+    assert all(
+        [
+            c1.name == c2.name and c1 != c2
+            for c1, c2 in zip(t1.columns, t2.columns)
+        ]
+    )
