+++ db/tests/test_columns.py
@@ -9,7 +9,9 @@ from sqlalchemy import (
     Sequence, Numeric, DateTime, func, UniqueConstraint,
 )
 from sqlalchemy.exc import IntegrityError
-from db import columns, tables, constants, constraints
+from db import columns, constants, constraints
+from db.tables import ddl as table_ddl
+from db.tables import utils as table_utils
 from db.types import email, alteration
 from db.types.base import get_db_type_name
 from db.tests.types import fixtures
@@ -38,11 +40,11 @@ def _rename_column(table, old_col_name, new_col_name, engine):
     """
     Renames the colum of a table and assert the change went through
     """
-    table_oid = tables.get_oid_from_table(table.name, table.schema, engine)
+    table_oid = table_utils.get_oid_from_table(table.name, table.schema, engine)
     column_index = columns.get_column_index_from_name(table_oid, old_col_name, engine)
     with engine.begin() as conn:
         columns.rename_column(table, column_index, engine, conn, new_col_name)
-    table = tables.reflect_table(table.name, table.schema, engine)
+    table = table_utils.reflect_table(table.name, table.schema, engine)
     assert new_col_name in table.columns
     assert old_col_name not in table.columns
     return table
@@ -256,7 +258,7 @@ def test_alter_column_chooses_wisely(column_dict, func_name, engine_with_schema)
     metadata = MetaData(bind=engine, schema=schema)
     table = Table(table_name, metadata, Column('col', String))
     table.create()
-    table_oid = tables.get_oid_from_table(table.name, table.schema, engine)
+    table_oid = table_utils.get_oid_from_table(table.name, table.schema, engine)
 
     with patch.object(columns, func_name) as mock_alterer:
         columns.alter_column(
@@ -283,8 +285,8 @@ def test_rename_column_foreign_keys(engine_with_schema):
     engine, schema = engine_with_schema
     table_name = "table_to_split"
     columns_list = [Column("Filler 1", Integer), Column("Filler 2", Integer)]
-    tables.create_mathesar_table(table_name, schema, columns_list, engine)
-    extracted, remainder, fk_name = tables.extract_columns_from_table(
+    table_ddl.create_mathesar_table(table_name, schema, columns_list, engine)
+    extracted, remainder, fk_name = table_ddl.extract_columns_from_table(
         table_name, ["Filler 1"], "Extracted", "Remainder", schema, engine
     )
     new_fk_name = "new_" + fk_name
@@ -300,7 +302,7 @@ def test_rename_column_sequence(engine_with_schema):
     new_col_name = "new_" + constants.ID
     engine, schema = engine_with_schema
     table_name = "table_with_columns"
-    table = tables.create_mathesar_table(table_name, schema, [], engine)
+    table = table_ddl.create_mathesar_table(table_name, schema, [], engine)
     with engine.begin() as conn:
         ins = table.insert()
         conn.execute(ins)
@@ -346,7 +348,7 @@ def test_get_column_index_from_name(engine_with_schema):
         Column(one_name, String),
     )
     table.create()
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     assert columns.get_column_index_from_name(table_oid, zero_name, engine) == 0
     assert columns.get_column_index_from_name(table_oid, one_name, engine) == 1
 
@@ -451,10 +453,10 @@ def test_create_column(engine_email_type, target_type):
         Column(initial_column_name, Integer),
     )
     table.create()
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     column_data = {"name": new_column_name, "type": target_type}
     created_col = columns.create_column(engine, table_oid, column_data)
-    altered_table = tables.reflect_table_from_oid(table_oid, engine)
+    altered_table = table_utils.reflect_table_from_oid(table_oid, engine)
     assert len(altered_table.columns) == 2
     assert created_col.name == new_column_name
     assert created_col.type.compile(engine.dialect) == input_output_type_map[target_type]
@@ -472,14 +474,14 @@ def test_create_column_options(engine_email_type, target_type):
         Column(initial_column_name, Integer),
     )
     table.create()
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     column_data = {
         "name": new_column_name,
         "type": target_type,
         "type_options": {"precision": 5, "scale": 3},
     }
     created_col = columns.create_column(engine, table_oid, column_data)
-    altered_table = tables.reflect_table_from_oid(table_oid, engine)
+    altered_table = table_utils.reflect_table_from_oid(table_oid, engine)
     assert len(altered_table.columns) == 2
     assert created_col.name == new_column_name
     assert created_col.plain_type == "NUMERIC"
@@ -498,7 +500,7 @@ def test_create_column_bad_options(engine_with_schema):
         Column(initial_column_name, Integer),
     )
     table.create()
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     column_data = {
         "name": new_column_name,
         "type": target_type,
@@ -532,7 +534,7 @@ def test_change_column_nullable_changes(engine_with_schema, nullable_tup):
             conn,
             nullable_tup[1],
         )
-    changed_table = tables.reflect_table(table_name, schema, engine)
+    changed_table = table_utils.reflect_table(table_name, schema, engine)
     changed_column = columns.get_mathesar_column_with_engine(
         changed_table.columns[0],
         engine
@@ -568,7 +570,7 @@ def test_change_column_nullable_with_data(engine_with_schema, nullable_tup):
             conn,
             nullable_tup[1],
         )
-    changed_table = tables.reflect_table(table_name, schema, engine)
+    changed_table = table_utils.reflect_table(table_name, schema, engine)
     changed_column = columns.get_mathesar_column_with_engine(
         changed_table.columns[0],
         engine
@@ -619,9 +621,9 @@ def test_drop_column_correct_column(engine_with_schema):
         Column(nontarget_column_name, String),
     )
     table.create()
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     columns.drop_column(table_oid, 0, engine)
-    altered_table = tables.reflect_table_from_oid(table_oid, engine)
+    altered_table = table_utils.reflect_table_from_oid(table_oid, engine)
     assert len(altered_table.columns) == 1
     assert nontarget_column_name in altered_table.columns
     assert target_column_name not in altered_table.columns
@@ -660,7 +662,7 @@ def test_get_column_default(engine_with_schema, filler, col_type):
         *cols
     )
     table.create()
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
 
     default = columns.get_column_default(table_oid, 0, engine)
     created_default = _get_default(engine, table)
@@ -685,7 +687,7 @@ def test_get_column_generated_default(engine_with_schema, col):
         col,
     )
     table.create()
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     default = columns.get_column_default(table_oid, 0, engine)
     created_default = _get_default(engine, table)
 
@@ -709,7 +711,7 @@ def test_column_default_create(engine_with_schema, col_type):
 
     with engine.begin() as conn:
         columns.set_column_default(table, 0, engine, conn, set_default)
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     default = columns.get_column_default(table_oid, 0, engine)
     created_default = _get_default(engine, table)
 
@@ -732,7 +734,7 @@ def test_column_default_update(engine_with_schema, col_type):
 
     with engine.begin() as conn:
         columns.set_column_default(table, 0, engine, conn, set_default)
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     default = columns.get_column_default(table_oid, 0, engine)
     created_default = _get_default(engine, table)
 
@@ -756,7 +758,7 @@ def test_column_default_delete(engine_with_schema, col_type):
 
     with engine.begin() as conn:
         columns.set_column_default(table, 0, engine, conn, None)
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     default = columns.get_column_default(table_oid, 0, engine)
     created_default = _get_default(engine, table)
 
@@ -773,7 +775,7 @@ duplicate_column_options = [
 
 
 def _check_duplicate_data(table_oid, engine, copy_data):
-    table = tables.reflect_table_from_oid(table_oid, engine)
+    table = table_utils.reflect_table_from_oid(table_oid, engine)
 
     with engine.begin() as conn:
         rows = conn.execute(table.select()).fetchall()
@@ -819,9 +821,9 @@ def test_duplicate_column_name(engine_with_schema):
         Column("Filler", Numeric)
     )
     table.create()
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     columns.duplicate_column(table_oid, 0, engine, new_col_name)
-    table = tables.reflect_table_from_oid(table_oid, engine)
+    table = table_utils.reflect_table_from_oid(table_oid, engine)
     assert new_col_name in table.c
 
 
@@ -835,7 +837,7 @@ def test_duplicate_column_single_unique(engine_with_schema, copy_data, copy_cons
     insert_data = [(1,), (2,), (3,)]
     _create_table(table_name, cols, insert_data, schema, engine)
 
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     columns.duplicate_column(
         table_oid, 0, engine, new_col_name, copy_data, copy_constraints
     )
@@ -861,7 +863,7 @@ def test_duplicate_column_multi_unique(engine_with_schema, copy_data, copy_const
     insert_data = [(1, 2), (2, 3), (3, 4)]
     _create_table(table_name, cols, insert_data, schema, engine)
 
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     columns.duplicate_column(
         table_oid, 0, engine, new_col_name, copy_data, copy_constraints
     )
@@ -886,7 +888,7 @@ def test_duplicate_column_nullable(
     insert_data = [(1,), (2,), (3,)]
     _create_table(table_name, cols, insert_data, schema, engine)
 
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     col = columns.duplicate_column(
         table_oid, 0, engine, new_col_name, copy_data, copy_constraints
     )
@@ -916,7 +918,7 @@ def test_duplicate_non_unique_constraint(engine_with_schema):
         for data in insert_data:
             conn.execute(table.insert().values(data))
 
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     col = columns.duplicate_column(table_oid, 0, engine, new_col_name)
 
     _check_duplicate_data(table_oid, engine, True)
@@ -933,7 +935,7 @@ def test_duplicate_column_default(engine_with_schema, copy_data, copy_constraint
     cols = [Column(target_column_name, Numeric, server_default=str(expt_default))]
     _create_table(table_name, cols, [], schema, engine)
 
-    table_oid = tables.get_oid_from_table(table_name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
     columns.duplicate_column(
         table_oid, 0, engine, new_col_name, copy_data, copy_constraints
     )
@@ -1004,10 +1006,10 @@ def _get_pizza_column_data():
 def test_batch_update_columns_no_changes(engine_email_type):
     engine, schema = engine_email_type
     table = _create_pizza_table(engine, schema)
-    table_oid = tables.get_oid_from_table(table.name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table.name, schema, engine)
 
     columns.batch_update_columns(table_oid, engine, _get_pizza_column_data())
-    updated_table = tables.reflect_table(table.name, schema, engine)
+    updated_table = table_utils.reflect_table(table.name, schema, engine)
 
     assert len(table.columns) == len(updated_table.columns)
     for index, column in enumerate(table.columns):
@@ -1019,14 +1021,14 @@ def test_batch_update_columns_no_changes(engine_email_type):
 def test_batch_update_column_names(engine_email_type):
     engine, schema = engine_email_type
     table = _create_pizza_table(engine, schema)
-    table_oid = tables.get_oid_from_table(table.name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table.name, schema, engine)
 
     column_data = _get_pizza_column_data()
     column_data[1]['name'] == 'Pizza Style'
     column_data[2]['name'] == 'Eaten Recently?'
 
     columns.batch_update_columns(table_oid, engine, column_data)
-    updated_table = tables.reflect_table(table.name, schema, engine)
+    updated_table = table_utils.reflect_table(table.name, schema, engine)
 
     assert len(table.columns) == len(updated_table.columns)
     for index, column in enumerate(table.columns):
@@ -1038,14 +1040,14 @@ def test_batch_update_column_names(engine_email_type):
 def test_batch_update_column_types(engine_email_type):
     engine, schema = engine_email_type
     table = _create_pizza_table(engine, schema)
-    table_oid = tables.get_oid_from_table(table.name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table.name, schema, engine)
 
     column_data = _get_pizza_column_data()
     column_data[0]['plain_type'] == 'INTEGER'
     column_data[2]['plain_type'] == 'BOOLEAN'
 
     columns.batch_update_columns(table_oid, engine, column_data)
-    updated_table = tables.reflect_table(table.name, schema, engine)
+    updated_table = table_utils.reflect_table(table.name, schema, engine)
 
     assert len(table.columns) == len(updated_table.columns)
     for index, column in enumerate(table.columns):
@@ -1057,7 +1059,7 @@ def test_batch_update_column_types(engine_email_type):
 def test_batch_update_column_names_and_types(engine_email_type):
     engine, schema = engine_email_type
     table = _create_pizza_table(engine, schema)
-    table_oid = tables.get_oid_from_table(table.name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table.name, schema, engine)
 
     column_data = _get_pizza_column_data()
     column_data[0]['name'] == 'Pizza ID'
@@ -1066,7 +1068,7 @@ def test_batch_update_column_names_and_types(engine_email_type):
     column_data[2]['plain_type'] == 'BOOLEAN'
 
     columns.batch_update_columns(table_oid, engine, column_data)
-    updated_table = tables.reflect_table(table.name, schema, engine)
+    updated_table = table_utils.reflect_table(table.name, schema, engine)
 
     assert len(table.columns) == len(updated_table.columns)
     for index, column in enumerate(table.columns):
@@ -1078,14 +1080,14 @@ def test_batch_update_column_names_and_types(engine_email_type):
 def test_batch_update_column_drop_columns(engine_email_type):
     engine, schema = engine_email_type
     table = _create_pizza_table(engine, schema)
-    table_oid = tables.get_oid_from_table(table.name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table.name, schema, engine)
 
     column_data = _get_pizza_column_data()
     column_data[0] = {}
     column_data[1] = {}
 
     columns.batch_update_columns(table_oid, engine, column_data)
-    updated_table = tables.reflect_table(table.name, schema, engine)
+    updated_table = table_utils.reflect_table(table.name, schema, engine)
 
     assert len(updated_table.columns) == len(table.columns) - 2
     for index, column in enumerate(updated_table.columns):
@@ -1097,7 +1099,7 @@ def test_batch_update_column_drop_columns(engine_email_type):
 def test_batch_update_column_all_operations(engine_email_type):
     engine, schema = engine_email_type
     table = _create_pizza_table(engine, schema)
-    table_oid = tables.get_oid_from_table(table.name, schema, engine)
+    table_oid = table_utils.get_oid_from_table(table.name, schema, engine)
 
     column_data = _get_pizza_column_data()
     column_data[0]['name'] = 'Pizza ID'
@@ -1107,7 +1109,7 @@ def test_batch_update_column_all_operations(engine_email_type):
     column_data[3] = {}
 
     columns.batch_update_columns(table_oid, engine, column_data)
-    updated_table = tables.reflect_table(table.name, schema, engine)
+    updated_table = table_utils.reflect_table(table.name, schema, engine)
 
     assert len(updated_table.columns) == len(table.columns) - 1
     for index, column in enumerate(updated_table.columns):
