+++ mathesar/api/viewsets/__init__.py
@@ -1,0 +1,7 @@
+from mathesar.api.viewsets.columns import ColumnViewSet  # noqa
+from mathesar.api.viewsets.constraints import ConstraintViewSet # noqa
+from mathesar.api.viewsets.data_files import DataFileViewSet # noqa
+from mathesar.api.viewsets.databases import DatabaseViewSet # noqa
+from mathesar.api.viewsets.records import RecordViewSet # noqa
+from mathesar.api.viewsets.schemas import SchemaViewSet # noqa
+from mathesar.api.viewsets.tables import TableViewSet # noqa
