+++ mathesar_ui/README.md
@@ -49,19 +49,19 @@ If you want to add or remove packages, or bascially run any npm action, **always
 You can connect to the container and open the ui folder by running:
 
 ```bash
-docker exec --user mathesar -it mathesar_service_1 /bin/bash
+docker exec -it mathesar_service_1 /bin/bash
 cd mathesar_ui
 ```
 
 and then perform any action from within it. Example:
 ```bash
-node@c273da65c52d:/code/mathesar_ui$ ls
+root@c273da65c52d:/code/mathesar_ui$ ls
 Dockerfile  jsconfig.json  package-lock.json  public  vite.config.js
 README.md   node_modules   package.json       src
 
-node@c273da65c52d:/code/mathesar_ui$ npm install <package>
+root@c273da65c52d:/code/mathesar_ui$ npm install <package>
 
-node@c273da65c52d:/code/mathesar_ui$ npm uninstall <package>
+root@c273da65c52d:/code/mathesar_ui$ npm uninstall <package>
 ```
 
 ### Components
@@ -71,14 +71,23 @@ For guidelines on component development, refer [README of components](https://gi
 You can start storybook in dev mode by connecting to the container and then running `npm run storybook`:
 
 ```bash
-docker exec --user mathesar -it mathesar_service_1 /bin/bash
+docker exec -it mathesar_service_1 /bin/bash
 
-node@c273da65c52d:/code/mathesar_ui$ cd mathesar_ui && npm run storybook
+root@c273da65c52d:/code/mathesar_ui$ cd mathesar_ui && npm run storybook
 ```
 
 This should start storybook at port 6006.
 
 To build storybook, run `npm run build-storybook`. The static storybook build will be present within __storybook-static__ folder.
+
+### Developing in Windows
+Hot module replacement does not work well with WSL when the project is present within a Windows filesystem, as mentioned in [this issue](https://github.com/microsoft/WSL/issues/4739).
+
+The simplest way to get this to work (and the one we advise) is to move the project to a Linux filesystem. This can be achieved by cloning the repo within the WSL shell in a Linux filesystem.
+
+If you have to work in the Windows filesystem, you could configure Vite to poll files to identify changes, as mentioned in [this issue](https://github.com/vitejs/vite/issues/1153#issuecomment-785467271) and in the [Vite documentation](https://vitejs.dev/config/#server-watch). However, this is a resource intensive process so we advise the previous option instead.
+
+[This issue](https://github.com/centerofci/mathesar/issues/570) keeps track of problems encountered by Mathesar developers using Windows for local development.
 
 ## Naming conventions
