+++ mathesar_ui/src/sections/table-view/actions-pane/ActionsPane.svelte
@@ -44,9 +44,12 @@
 
   function openConfirmation(){
     //if table.not empty call store to delete table else show confirmation
-    var ggg = getTable($currentDBName,$currentSchemaId);
-    var records = ggg.records;
-    console.log(get(records));
+    // var table = getTable($currentDBName,$currentSchemaId);
+    // var records = table.records;
+    // console.log(get(records));
+    const { activeTab }  = getTabsForSchema($currentDBName,$currentSchemaId);
+    const activeTabObj = get(activeTab)
+    console.log(activeTabObj.id);
   }
   
   async function tableDelete(){
