+++ run.sh
@@ -12,7 +12,7 @@
 # client. Hence, instead of using this script, the web server
 # can be directly started.
 
-cd mathesar_ui && npm install
+cd mathesar_ui && npm install --unsafe-perm
 
 # Run vite dev server only in dev mode
 if [[ "$MODE" == "DEVELOPMENT" || "$1" == "dev" ]]; then
